third_party/blink/web_tests/fast/multicol/orphans-relayout-expected.txt

Test if an element with orphans relayouts correctly. The red lines of text must be at the top of the blue rectangle. The green lines of text must be at the bottom of the blue rectangle.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS test Block 1 Line 1 is correct.
PASS test Block 2 Line 1 is correct.
PASS successfullyParsed is true

TEST COMPLETE
